+++ railties/lib/rails/application/default_middleware_stack.rb
@@ -47,7 +47,6 @@ module Rails
             end
           end
 
-          middleware.use ::Rack::Runtime
           middleware.use ::Rack::MethodOverride unless config.api_only
           middleware.use ::ActionDispatch::RequestId